Restructuring the repository for easier development of backend
