+++ src/core/domain/job_description.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class JobDescription(BaseModel):
+    company_name: st
+    title: str
+    description: str
