+++ src/core/ports/secondary/template_service.py
