+++ src/infrastructure/privacy/presidio_privacy_service.py
@@ -1,4 +1,4 @@
-from src.core.ports.privacy_filter import BasePrivacyFilter
+from src.core.ports.secondary.privacy_filter import BasePrivacyFilter
 from presidio_analyzer import AnalyzerEngine
 from presidio_anonymizer import AnonymizerEngine
 from presidio_anonymizer.entities import OperatorConfig
