+++ src/infrastructure/template/exceptions.py
