+++ src/infrastructure/template/templates/prompts/optimization/keyword_identification.j2
